# Conflicts:
#	app/uk/gov/hmrc/tradergoodsprofiles/connectors/RouterConnector.scala
#	app/uk/gov/hmrc/tradergoodsprofiles/controllers/RemoveRecordController.scala
#	app/uk/gov/hmrc/tradergoodsprofiles/models/requests/UpdateRecordRequest.scala
#	app/uk/gov/hmrc/tradergoodsprofiles/utils/ApplicationConstants.scala
#	app/uk/gov/hmrc/tradergoodsprofiles/utils/ValidationSupport.scala
#	it/test/uk/gov/hmrc/tradergoodsprofiles/controllers/RemoveRecordControllerIntegrationSpec.scala
#	test/uk/gov/hmrc/tradergoodsprofiles/connectors/RouterConnectorSpec.scala
#	test/uk/gov/hmrc/tradergoodsprofiles/controllers/CreateRecordControllerSpec.scala
#	test/uk/gov/hmrc/tradergoodsprofiles/controllers/RemoveRecordControllerSpec.scala

+++ app/uk/gov/hmrc/tradergoodsprofiles/models/responses/GetRecordResponse.scala
@@ -33,7 +33,7 @@ case class GetRecordResponse(
   countryOfOrigin: String,
   category: Int,
   assessments: Option[Seq[Assessment]],
-  supplementaryUnit: Option[Int],
+  supplementaryUnit: Option[BigDecimal],
   measurementUnit: Option[String],
   comcodeEffectiveFromDate: Instant,
   comcodeEffectiveToDate: Option[Instant],
@@ -121,7 +121,7 @@ object GetRecordResponse {
       val countryOfOrigin          = read[String]("countryOfOrigin")
       val category                 = read[Int]("category")
       val assessments              = readOption[Seq[Assessment]]("assessments")
-      val supplementaryUnit        = readOption[Int]("supplementaryUnit")
+      val supplementaryUnit        = readOption[BigDecimal]("supplementaryUnit")
       val measurementUnit          = readOption[String]("measurementUnit")
       val comcodeEffectiveFromDate = read[Instant]("comcodeEffectiveFromDate")
       val comcodeEffectiveToDate   = readOption[Instant]("comcodeEffectiveToDate")
